tests/wpt/tests/speculation-rules/prerender/session-history-navigation.https.html

<!DOCTYPE html>
<title>Test history.length</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="../resources/utils.js"></script>
<script src="resources/utils.js"></script>
<script src="resources/session-history-test-util.js"></script>

<body>
<script>
setup(() => assertSpeculationRulesIsSupported());

const uid = token();

// We test only a fragment-navigation because other kinds of the main frame navigation
// in prerender cancels prerendering.
promise_test(async () => {
  assert_equals(
    await runTestInPrerender("testSyntheticAnchorClickInPrerender", uid),
    "Passed"
  );
}, "Synthetic anchor click navigates independently with replacement in a prerender");
</script>
